Fixup: Actually use all 48 bits for the position bitmask.

+++ code/features-search/result-ranking/src/test/java/nu/marginalia/ranking/ResultValuatorTest.java
@@ -84,7 +84,7 @@ class ResultValuatorTest {
 
     private long wordMetadata(int tfIdf, Set<Integer> positions, Set<WordFlags> wordFlags) {
         long posBits = positions.stream()
-                .mapToLong(i -> (int)((1L << i) & 0xFFFF_FFFF_FFFFL))
+                .mapToLong(i -> ((1L << i) & 0xFFFF_FFFF_FFFFL))
                 .reduce((a,b) -> a|b)
                 .orElse(0L);
